+++ src/client/lib/index.ts
@@ -1,0 +1,3 @@
+/* eslint-disable import/prefer-default-export */
+export const supportsClipboard = () =>
+  navigator.clipboard.readText !== undefined;