+++ src/components/forms/GSScriptOptionsField.jsx
@@ -36,12 +36,7 @@ class GSScriptOptionsField extends GSFormField {
     event.stopPropagation();
     event.preventDefault();
 
-    this.setState(
-      { scriptTarget: scriptVersion, scriptDraft: scriptVersion },
-      () => {
-        if (this.inputRef) this.inputRef.focus();
-      }
-    );
+    this.setState({ scriptTarget: scriptVersion, scriptDraft: scriptVersion });
   };
 
   createDeleteHandler = (scriptVersion) => () => {
@@ -145,6 +140,7 @@ class GSScriptOptionsField extends GSFormField {
             name={name}
             scriptText={scriptDraft}
             scriptFields={scriptFields}
+            receiveFocus
             expandable
             onChange={(val) => this.setState({ scriptDraft: val.trim() })}
           />
@@ -174,8 +170,8 @@ class GSScriptOptionsField extends GSFormField {
       "multiLine",
       "name",
       "data-test",
-      "onBlur",
-      "onChange"
+      "onBlur"
+      // We manage onChange ourselves so don't pass it though
     ]);
 
     const canDelete = scriptVersions.length > 1;